refactor: split non canonical signature task to individual class
