+++ test/BasicKeyPairTest.template.ts
@@ -166,43 +166,6 @@ export const BasicKeyPairTester = (keyPairClass: KeyPairClass, deterministicPriv
             // Assert:
             expect(isVerified).to.equal(false);
         });
-
-        it('verify non canonical signature', () => {
-            /**
-             * @param scalar - scalar
-             */
-            function scalarAddGroupOrder(scalar: Uint8Array): void {
-                // 2^252 + 27742317777372353535851937790883648493, little endian
-                const Group_Order = [
-                    0xed, 0xd3, 0xf5, 0x5c, 0x1a, 0x63, 0x12, 0x58, 0xd6, 0x9c, 0xf7, 0xa2, 0xde, 0xf9, 0xde, 0x14, 0x00, 0x00, 0x00, 0x00,
-                    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10,
-                ];
-
-                let r = 0;
-                for (let i = 0; i < scalar.length; ++i) {
-                    const t = scalar[i] + Group_Order[i];
-                    scalar[i] += Group_Order[i] + r;
-                    r = (t >> 8) & 0xff;
-                }
-            }
-
-            // Arrange:
-            const keyPair: KeyPair = new keyPairClass(deterministicPrivateKey);
-            const payload = new Uint8Array([1, 2, 3, 4, 5, 6, 7, 8, 9, 11]);
-            const canonicalSignature = keyPair.sign(payload);
-
-            // this is signature with group order added to 'encodedS' part of signature
-            const nonCanonicalSignature = canonicalSignature.slice();
-            scalarAddGroupOrder(nonCanonicalSignature.subarray(32));
-
-            // Act:
-            const isCanonicalVerified = keyPair.verify(payload, canonicalSignature);
-            const isNonCanonicalVerified = keyPair.verify(payload, nonCanonicalSignature);
-
-            // Assert:
-            expect(isCanonicalVerified).to.equal(true);
-            expect(isNonCanonicalVerified).to.equal(false);
-        });
     });
 
     describe('generate', () => {
